+++ source/TestCommon/source/FunctionApp.TestCommon.Tests/Fixtures/ServiceBusListenerMockFixture.cs
@@ -12,14 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.FunctionApp.TestCommon.Configuration;
 using Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ListenerMock;
 using Energinet.DataHub.Core.TestCommon.Diagnostics;
 
 namespace Energinet.DataHub.Core.FunctionApp.TestCommon.Tests.Fixtures;
 
 /// <summary>
-/// This fixtures ensures we reuse <see cref="ConnectionString"/>
+/// This fixtures ensures we reuse <see cref="FullyQualifiedNamespace"/>
 /// so we only have to retrieve an access token and values in Key Vault one time.
 ///
 /// When testing the <see cref="ServiceBusListenerMock"/> we must create new queues/topics
@@ -31,11 +30,10 @@ public class ServiceBusListenerMockFixture
     public ServiceBusListenerMockFixture()
     {
         TestLogger = new TestDiagnosticsLogger();
-
-        ConnectionString = SingletonIntegrationTestConfiguration.Instance.ServiceBusConnectionString;
+        FullyQualifiedNamespace = SingletonIntegrationTestConfiguration.Instance.ServiceBusFullyQualifiedNamespace;
     }
 
     public ITestDiagnosticsLogger TestLogger { get; }
 
-    public string ConnectionString { get; }
+    public string FullyQualifiedNamespace { get; }
 }